requirements.txt
streamlit
openai
python-dotenv
faiss-cpu
PyPDF2
numpy